Move SNS and SQS to product service
