+++ packages/import-service/.env.example
@@ -4,11 +4,3 @@ PRODUCTS_FILE_EXTENSION=.csv
 UPLOAD_BUCKET=
 UPLOADED_FOLDER=
 PARSED_FOLDER=
-SUBSCRIPTION_ENDPOINT_1=
-
-# DB connection
-PGUSER=
-PGHOST=
-PGPASSWORD=
-PGDATABASE=
-PGPORT=
