+++ packages/product/.env.example
@@ -3,3 +3,6 @@ PGHOST=
 PGPASSWORD=
 PGDATABASE=
 PGPORT=
+
+SUBSCRIPTION_ENDPOINT_1=
+SUBSCRIPTION_ENDPOINT_2=
