+++ packages/product/handlers/__mocks__/aws-sdk.js
@@ -1,9 +1,15 @@
-export const getObject = jest.fn(() => Promise.resolve());
-export const putObject = jest.fn(() => Promise.resolve());
-export const getSignedUrl = jest.fn(() => Promise.resolve());
-export const deleteObject = jest.fn(() => Promise.resolve());
-export const copyObject = jest.fn(() => Promise.resolve());
+import {
+  copyObject,
+  deleteObject,
+  getObject,
+  getSignedUrl,
+  putObject,
+} from '../../../import-service/handlers/__mocks__/aws-sdk';
+
+export let publish;
+
 export const sendMessage = jest.fn(() => Promise.resolve());
+export const promise = jest.fn(() => Promise.resolve());
 
 const AWS = {
   S3: class S3 {
@@ -15,10 +21,17 @@ const AWS = {
       this.copyObject = copyObject;
     }
   },
-
   SQS: class SQS {
     constructor() {
       this.sendMessage = sendMessage;
+    }
+  },
+
+  SNS: class SNS {
+    constructor() {
+      publish = jest.fn().mockReturnValue(this);
+      this.publish = publish;
+      this.promise = promise;
     }
   },
 };
