+++ packages/product/serverless.yml
@@ -18,6 +18,45 @@ provider:
   stage: dev
   profile: epam-aws-training-account
 
+  environment:
+    CREATE_PRODUCT_TOPIC:
+      Ref: createProductTopic
+
+  iamRoleStatements:
+    - Effect: "Allow"
+      Action:
+        - "sns:*"
+      Resource:
+        Ref: createProductTopic
+
+resources:
+  Resources:
+    createProductTopic:
+      Type: "AWS::SNS::Topic"
+      Properties:
+        TopicName: createProductTopic
+    createProductSubscription:
+      Type: "AWS::SNS::Subscription"
+      Properties:
+        Endpoint: ${env:SUBSCRIPTION_ENDPOINT_1}
+        Protocol: email
+        TopicArn:
+          Ref: createProductTopic
+        FilterPolicy: {
+          "title": ["Classic tour of Minsk"]
+        }
+
+    createExpensiveProductSubscription:
+      Type: "AWS::SNS::Subscription"
+      Properties:
+        Endpoint: ${env:SUBSCRIPTION_ENDPOINT_2}
+        Protocol: email
+        TopicArn:
+          Ref: createProductTopic
+        FilterPolicy: {
+          "title": [{"anything-but":"Classic tour of Minsk"}]
+        }
+
 functions:
   getProducts:
     handler: ${self:custom.distFolder}/get-products.getProducts
@@ -40,3 +79,10 @@ functions:
           method: post
           path: /products
           cors: true
+  catalogBatchProcess:
+    handler: ${self:custom.distFolder}/catalog-batch-process.catalogBatchProcess
+    events:
+      - sqs:
+          batchSize: 5
+          arn:
+            !ImportValue myCatalogItemsQueueArn
