+++ packages/product/webpack.config.js
@@ -2,13 +2,17 @@ const path = require('path');
 const glob = require('glob');
 const baseConfig = require('../../webpack-base.config');
 
-const entries = glob.sync('./handlers/*.js').reduce(
-  (x, y) =>
-    Object.assign(x, {
-      [y]: y,
-    }),
-  {},
-);
+const entries = glob
+  .sync('./handlers/*.js', {
+    ignore: './handlers/*.test.js',
+  })
+  .reduce(
+    (x, y) =>
+      Object.assign(x, {
+        [y]: y,
+      }),
+    {},
+  );
 
 module.exports = {
   ...baseConfig,
